+++ crates/dictymus-core/src/lib.rs
@@ -2,4 +2,5 @@ pub mod config;
 pub mod dictionary;
 pub mod language;
 pub mod normalize;
+pub mod testing;
 pub mod transliterate;