+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Opening a dictionary now places keyboard focus in the search field instead
+  of the tab bar
+
 ## [0.1.0] - 2026-08-14
 
 ### Added